Refactor to leverage new libraries
* This leverages planetscale/psdb for the grpc library fundamentals
* Uses new psdbconnect service definition out of edge-gateway

The result of these changes is that psdbconnect is split out of `psdb`
as a service namespace, and better allows us to make planetscale/psdb
open source soon without exposing all of `Connect/Sync` service yet
until we solidify these APIs.

The issue is trying to avoid exposing Vitess types and Vitess protos as
a part of our public API and not have a dependency on vitess within
planetscale/psdb.