Update to latest Go & dependencies and improve build
This updates to the latest Go 1.20.1 which resolves a bunch of
Dependabot alerts as well and keeps things updated.

It also updates the dependencies as well and improves the build system a
bit to use `-trimpath` so we have better reproduceable builds and don't
include random build time paths in the final binary.