This PR makes a few changes to the Sync API used by the various ELT tools we build & ship, i.e. Airbyte, Fivetran and Stitch connectors.
Introduce a "columns" array in the SyncRequest that we'll use to filter out which columns are returned from vstream in a Sync request.
Introduce "IncludeInserts" to send any new rows from Vstream to the client.
Introduce "IncludeUpdates" to send any updated rows from Vstream to the client.
Introduce "IncludeDeletes" to send any deleted rows from Vstream to the client.