Don't use a client side read timeout
The server already has it's own timeout stream, and having this client
side causes us to prematurely close the connection potentially
mid-stream without the ability to gracefully stop.

The max duration isn't a client concern, and is purely a server concern.
Removing the timeout here allows the server to increase or decrease it's
timeouts or have any different behaviors without coordination with the
client.