Improve the type mapping
This improves the type mapping based on the documentation and refactors
the code to be easier to follow. It's now again a single switch
statement without duplicated logic in places.

We treat all integer type like types in MySQL as integer. A `bigint`
still fits in 64 bits and is better annotated as an Airbyte integer
type.

It also adds handling for `time` and `float` as well which have better
matching types too. Similar for the MySQL `timestamp` type which can be
treated like a `datetime`.