add option to use gtid position together with last known pk (#140)
Currently when airbyte-source syncs it wipes out `gtid` information from
the shard gtid before passing it along in the VStream request. This
means that, upon retry, it's possible for the connector to lose rows
that are inserted between the `gtid` of the last `vgtid` event it
received in one attempt and the `@@gtid_executed` of the next sync
attempt.

This new option changes the current behavior so that `gtid` is always
passed along in the VStream request, if it is available. This makes for
a safer retry. However, it does mean that it is possible for the
connector to received duplicates of the same row.

How it will look in Airbyte UI after upgrading to a version with this
change.

<img width="849" alt="image"
src="https://github.com/user-attachments/assets/5d699e82-3e37-4459-a833-365f717cea15"
/>

Signed-off-by: Max Englander <max@planetscale.com>