add option for specifying timeout in minutes and gtid position together with last known pk (#139)
Add two new options.

## Specifying timeout in minutes

Make the current hardcoded 5 minute read duration configurable. Keep 5
minutes as the default and minimum. This is how it will look on
connectors that upgrade to a version containing this change:

<img width="846" alt="image"
src="https://github.com/user-attachments/assets/d6b54b3f-af9d-4132-a5fb-92031073c501"
/>

## GTID position with last known PK

Currently when airbyte-source syncs it wipes out `gtid` information from
the shard gtid before passing it along in the VStream request. This
means that, upon retry, it's possible for the connector to lose rows
that are inserted between the `gtid` of the last `vgtid` event it
received in one attempt and the `@@gtid_executed` of the next sync
attempt.

This new option changes the current behavior so that `gtid` is always
passed along in the VStream request, if it is available. This makes for
a safer retry. However, it does mean that it is possible for the
connector to received duplicates of the same row.

How it will look in Airbyte UI after upgrading to a version with this
change.

<img width="849" alt="image"
src="https://github.com/user-attachments/assets/5d699e82-3e37-4459-a833-365f717cea15"
/>

---------

Signed-off-by: Max Englander <max@planetscale.com>